+++ src/commands/ext-dev-register.ts
@@ -1,5 +1,6 @@
 import * as clc from "cli-color";
-import * as marked from "marked";
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-var-requires
+const { marked } = require("marked");
 
 import { Command } from "../command";
 import { registerPublisherProfile } from "../extensions/extensionsApi";
@@ -35,7 +36,7 @@ export default new Command("ext:dev:register")
     });
     try {
       await registerPublisherProfile(projectId, publisherId);
-    } catch (err) {
+    } catch (err: any) {
       if (err.status === 409) {
         const error =
           `Couldn't register the publisher ID '${clc.bold(publisherId)}' to the project '${clc.bold(
@@ -52,8 +53,7 @@ export default new Command("ext:dev:register")
       throw new FirebaseError(
         `Failed to register publisher ID ${clc.bold(publisherId)} for project ${clc.bold(
           projectId
-        )}: ${err.message}`,
-        { exit: 1 }
+        )}: ${err.message}`
       );
     }
     return utils.logLabeledSuccess(