+++ src/deploy/functions/runtimes/node/index.ts
@@ -6,12 +6,9 @@ import * as spawn from "cross-spawn";
 import fetch from "node-fetch";
 
 import { FirebaseError } from "../../../../error";
-import { Options } from "../../../../options";
 import { getRuntimeChoice } from "./parseRuntimeAndValidateSDK";
-import { needProjectId } from "../../../../projectUtils";
 import { logger } from "../../../../logger";
 import { previews } from "../../../../previews";
-import * as args from "../../args";
 import * as backend from "../../backend";
 import * as runtimes from "..";
 import * as validate from "./validate";
@@ -19,16 +16,10 @@ import * as versioning from "./versioning";
 import * as parseTriggers from "./parseTriggers";
 import * as discovery from "../discovery";
 
-/**
- *
- */
 export async function tryCreateDelegate(
-  context: args.Context,
-  options: Options
+  context: runtimes.DelegateContext
 ): Promise<Delegate | undefined> {
-  const projectRelativeSourceDir = options.config.get("functions.source") as string;
-  const sourceDir = options.config.path(projectRelativeSourceDir);
-  const packageJsonPath = path.join(sourceDir, "package.json");
+  const packageJsonPath = path.join(context.sourceDir, "package.json");
 
   if (!(await promisify(fs.exists)(packageJsonPath))) {
     logger.debug("Customer code is not Node");
@@ -36,10 +27,9 @@ export async function tryCreateDelegate(
   }
 
   // Check what runtime to use, first in firebase.json, then in 'engines' field.
-  let runtime = (options.config.get("functions.runtime") as runtimes.Runtime) || "";
   // TODO: This method loads the Functions SDK version which is then manually loaded elsewhere.
   // We should find a way to refactor this code so we're not repeatedly invoking node.
-  runtime = getRuntimeChoice(sourceDir, runtime);
+  const runtime = getRuntimeChoice(context.sourceDir, context.runtime);
 
   if (!runtime.startsWith("nodejs")) {
     logger.debug(
@@ -48,7 +38,7 @@ export async function tryCreateDelegate(
     throw new FirebaseError(`Unexpected runtime ${runtime}`);
   }
 
-  return new Delegate(needProjectId(options), options.config.projectDir, sourceDir, runtime);
+  return new Delegate(context.projectId, context.projectDir, context.sourceDir, runtime);
 }
 
 // TODO(inlined): Consider moving contents in parseRuntimeAndValidateSDK and validate around.