+++ src/emulator/auth/server.ts
@@ -271,6 +271,12 @@ export async function createApp(
         // TODO
         return true;
       },
+      byte() {
+        // Disable the "byte" format validation to allow stuffing arbitary
+        // strings in passwordHash etc. Needed because the emulator generates
+        // non-base64 hash strings like "fakeHash:salt=foo:password=bar".
+        return true;
+      },
     },
     plugins: [
       {
@@ -507,12 +513,14 @@ function toExegesisController(
 
 function wrapValidateBody(pluginContext: ExegesisPluginContext): void {
   // Apply fixes to body for Google REST API mapping compatibility.
-  const op = ((pluginContext as unknown) as {
-    _operation: {
-      validateBody?: ValidatorFunction;
-      _authEmulatorValidateBodyWrapped?: true;
-    };
-  })._operation;
+  const op = (
+    pluginContext as unknown as {
+      _operation: {
+        validateBody?: ValidatorFunction;
+        _authEmulatorValidateBodyWrapped?: true;
+      };
+    }
+  )._operation;
   if (op.validateBody && !op._authEmulatorValidateBodyWrapped) {
     const validateBody = op.validateBody.bind(op);
     op.validateBody = (body) => {