+++ src/emulator/functionsEmulatorShell.ts
@@ -1,5 +1,5 @@
 import * as uuid from "uuid";
-import { FunctionsEmulator } from "./functionsEmulator";
+import { EmulatableBackend, FunctionsEmulator } from "./functionsEmulator";
 import {
   EmulatedTriggerDefinition,
   getSignatureType,
@@ -19,7 +19,7 @@ export class FunctionsEmulatorShell implements FunctionsShellController {
   emulatedFunctions: string[];
   urls: { [name: string]: string } = {};
 
-  constructor(private emu: FunctionsEmulator) {
+  constructor(private emu: FunctionsEmulator, private backend: EmulatableBackend) {
     this.triggers = emu.getTriggerDefinitions();
     this.emulatedFunctions = this.triggers.map((t) => t.id);
 
@@ -68,7 +68,13 @@ export class FunctionsEmulatorShell implements FunctionsShellController {
       data,
     };
 
-    this.emu.startFunctionRuntime(trigger.id, trigger.name, getSignatureType(trigger), proto);
+    this.emu.startFunctionRuntime(
+      this.backend,
+      trigger.id,
+      trigger.name,
+      getSignatureType(trigger),
+      proto
+    );
   }
 
   private getTrigger(name: string): EmulatedTriggerDefinition {