+++ src/emulator/loggingEmulator.ts
@@ -119,7 +119,7 @@ class WebSocketTransport extends TransportStream {
           try {
             bundle.data = { ...bundle.data, ...JSON.parse(value) };
             return null;
-          } catch (err) {
+          } catch (err: any) {
             // If the value isn't JSONable, just treat it like a string
             return value;
           }