+++ src/extensions/updateHelper.ts
@@ -1,6 +1,7 @@
 import * as clc from "cli-color";
 import * as semver from "semver";
-import * as marked from "marked";
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-var-requires
+const { marked } = require("marked");
 
 import { FirebaseError } from "../error";
 import { logger } from "../logger";
@@ -171,7 +172,7 @@ export async function updateFromLocalSource(
   let source;
   try {
     source = await createSourceFromLocation(projectId, localSource);
-  } catch (err) {
+  } catch (err: any) {
     throw new FirebaseError(invalidSourceErrMsgTemplate(instanceId, localSource));
   }
   utils.logLabeledBullet(
@@ -201,7 +202,7 @@ export async function updateFromUrlSource(
   let source;
   try {
     source = await createSourceFromLocation(projectId, urlSource);
-  } catch (err) {
+  } catch (err: any) {
     throw new FirebaseError(invalidSourceErrMsgTemplate(instanceId, urlSource));
   }
   utils.logLabeledBullet(
@@ -233,7 +234,7 @@ export async function updateToVersionFromPublisherSource(
   const extension = await extensionsApi.getExtension(extensionRef);
   try {
     source = await extensionsApi.getExtensionVersion(extVersionRef);
-  } catch (err) {
+  } catch (err: any) {
     throw new FirebaseError(
       `Could not find source '${clc.bold(extVersionRef)}' because (${clc.bold(
         version
@@ -245,7 +246,7 @@ export async function updateToVersionFromPublisherSource(
   let registryEntry;
   try {
     registryEntry = await resolveSource.resolveRegistryEntry(existingSpec.name);
-  } catch (err) {
+  } catch (err: any) {
     logger.debug(`Unable to fetch registry.json entry for ${existingSpec.name}`);
   }