+++ src/gcp/firedata.ts
@@ -1,4 +1,5 @@
-import * as api from "../api";
+import { firedataOrigin } from "../api";
+import { Client } from "../apiv2";
 import { logger } from "../logger";
 import * as utils from "../utils";
 
@@ -25,10 +26,13 @@ function _handleErrorResponse(response: any): any {
  * @return the list of databases.
  */
 export async function listDatabaseInstances(projectNumber: string): Promise<DatabaseInstance[]> {
-  const response = await api.request("GET", `/v1/projects/${projectNumber}/databases`, {
-    auth: true,
-    origin: api.firedataOrigin,
-  });
+  const client = new Client({ urlPrefix: firedataOrigin, apiVersion: "v1" });
+  const response = await client.get<{ instance: DatabaseInstance[] }>(
+    `/projects/${projectNumber}/databases`,
+    {
+      resolveOnHTTPError: true,
+    }
+  );
   if (response.status === 200) {
     return response.body.instance;
   }