+++ src/hosting/proxy.ts
@@ -40,7 +40,7 @@ function makeVary(vary: string | null = ""): string {
  * the Firebase Hosting origin.
  */
 export function proxyRequestHandler(url: string, rewriteIdentifier: string): RequestHandler {
-  return async (req: IncomingMessage, res: ServerResponse, next: () => void): Promise<void> => {
+  return async (req: IncomingMessage, res: ServerResponse, next: () => void): Promise<unknown> => {
     logger.info(`[hosting] Rewriting ${req.url} to ${url} for ${rewriteIdentifier}`);
     // Extract the __session cookie from headers to forward it to the
     // functions cookie is not a string[].
@@ -101,7 +101,7 @@ export function proxyRequestHandler(url: string, rewriteIdentifier: string): Req
         timeout: 60000,
         compress: false,
       });
-    } catch (err) {
+    } catch (err: any) {
       const isAbortError =
         err instanceof FirebaseError && err.original?.name.includes("AbortError");
       const isTimeoutError =
@@ -159,7 +159,7 @@ export function proxyRequestHandler(url: string, rewriteIdentifier: string): Req
           const unborkedLocation = location.replace(locationURL.origin, "");
           proxyRes.response.headers.set("location", unborkedLocation);
         }
-      } catch (e) {
+      } catch (e: any) {
         logger.debug(
           `[hosting] had trouble parsing location header, but this may be okay: "${location}"`
         );