+++ src/management/projects.ts
@@ -5,8 +5,7 @@ import * as ora from "ora";
 import { Client } from "../apiv2";
 import { FirebaseError } from "../error";
 import { pollOperation } from "../operation-poller";
-import { promptOnce } from "../prompt";
-import { Question } from "inquirer";
+import { Question, promptOnce } from "../prompt";
 import * as api from "../api";
 import { logger } from "../logger";
 import * as utils from "../utils";
@@ -84,7 +83,7 @@ export async function createFirebaseProjectAndLog(
   try {
     await createCloudProject(projectId, options);
     spinner.succeed();
-  } catch (err) {
+  } catch (err: any) {
     spinner.fail();
     throw err;
   }
@@ -100,7 +99,7 @@ export async function addFirebaseToCloudProjectAndLog(
 
   try {
     projectInfo = await addFirebaseToCloudProject(projectId);
-  } catch (err) {
+  } catch (err: any) {
     spinner.fail();
     throw err;
   }
@@ -271,7 +270,7 @@ export async function createCloudProject(
       operationResourceName: response.body.name /* LRO resource name */,
     });
     return projectInfo;
-  } catch (err) {
+  } catch (err: any) {
     if (err.status === 409) {
       throw new FirebaseError(
         `Failed to create project because there is already a project with ID ${clc.bold(
@@ -312,7 +311,7 @@ export async function addFirebaseToCloudProject(
       operationResourceName: response.body.name /* LRO resource name */,
     });
     return projectInfo;
-  } catch (err) {
+  } catch (err: any) {
     logger.debug(err.message);
     throw new FirebaseError(
       "Failed to add Firebase to Google Cloud Platform project. See firebase-debug.log for more info.",
@@ -366,7 +365,7 @@ export async function getFirebaseProjectPage(
       pageSize,
       pageToken,
     });
-  } catch (err) {
+  } catch (err: any) {
     logger.debug(err.message);
     throw new FirebaseError(
       "Failed to list Firebase projects. See firebase-debug.log for more info.",
@@ -391,7 +390,7 @@ export async function getAvailableCloudProjectPage(
       pageSize,
       pageToken,
     });
-  } catch (err) {
+  } catch (err: any) {
     logger.debug(err.message);
     throw new FirebaseError(
       "Failed to list available Google Cloud Platform projects. See firebase-debug.log for more info.",
@@ -432,7 +431,7 @@ export async function getFirebaseProject(projectId: string): Promise<FirebasePro
       timeout: TIMEOUT_MILLIS,
     });
     return res.body;
-  } catch (err) {
+  } catch (err: any) {
     logger.debug(err.message);
     throw new FirebaseError(
       `Failed to get Firebase project ${projectId}. ` +