+++ src/requireDatabaseInstance.ts
@@ -21,7 +21,7 @@ export async function requireDatabaseInstance(options: any): Promise<void> {
   let instance;
   try {
     instance = await getDefaultDatabaseInstance(options);
-  } catch (err) {
+  } catch (err: any) {
     throw new FirebaseError(`Failed to get details for project: ${options.project}.`, {
       original: err,
     });